Implement coin listing

+++ app/src/main/java/com/meliksahcakir/cointracker/data/CoinRepository.kt
@@ -1,6 +1,8 @@
 package com.meliksahcakir.cointracker.data
 
 import com.meliksahcakir.androidutils.Result
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 
 class CoinRepository(
     private val coinDao: CoinDao,
@@ -8,18 +10,23 @@ class CoinRepository(
     private val apiService: ApiService
 ) {
     private var coinInfoListAvailable = false
+    private var fetchingCoins = false
+    private var fetchingCoinInfoList = false
 
-    suspend fun fetchCoinInfoList() {
-        try {
-            val list = apiService.getCoinInfoList()
-            if (list.isNotEmpty()) {
-                coinDao.deleteAllCoinInfo()
-                coinDao.insertAllCoinInfo(*list.toTypedArray())
-                coinInfoListAvailable = true
+    suspend fun fetchCoinInfoList() = withContext(Dispatchers.IO) {
+        if (fetchingCoinInfoList) {
+            try {
+                val list = apiService.getCoinInfoList()
+                if (list.isNotEmpty()) {
+                    coinDao.deleteAllCoinInfo()
+                    coinDao.insertAllCoinInfo(*list.toTypedArray())
+                    coinInfoListAvailable = true
+                }
+            } catch (e: Exception) {
+                e.printStackTrace()
             }
-        } catch (e: Exception) {
-            e.printStackTrace()
         }
+        fetchingCoinInfoList = true
     }
 
     suspend fun searchCoins(query: String): Result<List<CoinInfo>> {
@@ -39,18 +46,25 @@ class CoinRepository(
     suspend fun fetchCoins(
         size: Int = ApiService.DEFAULT_PAGE_SIZE,
         page: Int = ApiService.DEFAULT_PAGE
-    ) {
-        try {
+    ): Result<Boolean> {
+        if (fetchingCoins) return Result.Success(false)
+        val result = try {
             val list = apiService.getCoinsMarketData(pageSize = size, page = page)
+                .map { it.mapToLocalModel() }
             coinDao.insertAllCoins(*list.toTypedArray())
+            Result.Success(list.isNotEmpty())
         } catch (e: Exception) {
             e.printStackTrace()
+            Result.Error(e)
         }
+        fetchingCoins = false
+        return result
     }
 
     suspend fun fetchFavoriteCoins(favorites: List<String>): Result<List<Coin>> {
         return try {
             val list = apiService.getCoinsMarketData(ids = favorites.joinToString(","))
+                .map { it.mapToLocalModel() }
             Result.Success(list)
         } catch (e: Exception) {
             Result.Error(e)
@@ -61,7 +75,7 @@ class CoinRepository(
         var error = false
         var details: CoinDetails? = null
         try {
-            details = apiService.getCoinDetails(id)
+            details = apiService.getCoinDetails(id).mapToLocalModel()
         } catch (e: Exception) {
             error = true
         }
